requirements.txt
edge-tts>=6.1.0
python-pptx>=0.6.23
pywin32>=306
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
